Bisect interface -- visuals only
Added a column which indicates goodness, badness, or untested for each
commit. No bisect calls are actually made yet, just the visual interface.
I think this can be modified so that when the normal git log calls are made
a git bisect log call can be used to automatically detect bisection and
turn on the visuals.